// sdk/src/realm/generated/models/WorldLorebookListDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WorldLorebookDetailDto } from './WorldLorebookDetailDto';
export type WorldLorebookListDto = {
    items: Array<WorldLorebookDetailDto>;
    worldId: string;
};